fix: add plugin host session cleanup seam (clawdbot-d02.1.9.1.38)
