+++ src/plugins/host-hook-cleanup.session-store.test.ts
@@ -101,4 +101,85 @@ describe("plugin host cleanup session stores", () => {
       },
     });
   });
+
+  it("clears plugin-owned session state across resolved stores without touching unrelated rows", async () => {
+    stateDir = await fs.mkdtemp(
+      path.join(resolvePreferredOpenClawTmpDir(), "openclaw-host-cleanup-multistore-"),
+    );
+    previousStateDir = process.env.OPENCLAW_STATE_DIR;
+    process.env.OPENCLAW_STATE_DIR = stateDir;
+    const firstStorePath = path.join(stateDir, "agent-a", "sessions.json");
+    const secondStorePath = path.join(stateDir, "agent-b", "sessions.json");
+    const beforeUpdatedAt = 100;
+    const unrelatedUpdatedAt = Date.now();
+    const firstEntry: SessionEntry = {
+      sessionId: "shared-session",
+      updatedAt: beforeUpdatedAt,
+      pluginExtensions: {
+        cleanup: { state: { active: true } },
+        other: { state: { preserved: true } },
+      },
+      pluginNextTurnInjections: {
+        cleanup: [
+          {
+            id: "remove",
+            pluginId: "cleanup",
+            text: "remove",
+            placement: "append_context",
+            createdAt: beforeUpdatedAt,
+          },
+        ],
+      },
+    };
+    const secondEntry: SessionEntry = {
+      sessionId: "shared-session",
+      updatedAt: beforeUpdatedAt,
+      pluginExtensions: {
+        cleanup: { state: { active: true } },
+      },
+    };
+    const unrelatedEntry: SessionEntry = {
+      sessionId: "unrelated-session",
+      updatedAt: unrelatedUpdatedAt,
+      pluginExtensions: {
+        cleanup: { state: { keep: true } },
+      },
+    };
+    await saveSessionStore(
+      firstStorePath,
+      {
+        "agent:a:main": firstEntry,
+        "agent:a:unrelated": unrelatedEntry,
+      },
+      { skipMaintenance: true },
+    );
+    await saveSessionStore(
+      secondStorePath,
+      {
+        "agent:b:other": secondEntry,
+      },
+      { skipMaintenance: true },
+    );
+
+    const result = await runPluginHostCleanup({
+      cfg: { session: { store: firstStorePath } },
+      registry: createEmptyPluginRegistry(),
+      pluginId: "cleanup",
+      reason: "disable",
+      sessionKey: "shared-session",
+      sessionStorePaths: [firstStorePath, secondStorePath],
+    });
+
+    expect(result).toEqual({ cleanupCount: 2, failures: [] });
+    const firstStore = loadSessionStore(firstStorePath, { skipCache: true });
+    const secondStore = loadSessionStore(secondStorePath, { skipCache: true });
+    expect(firstStore["agent:a:main"]?.pluginExtensions).toEqual({
+      other: { state: { preserved: true } },
+    });
+    expect(firstStore["agent:a:main"]?.pluginNextTurnInjections).toBeUndefined();
+    expect(firstStore["agent:a:main"]?.updatedAt).toBeGreaterThan(beforeUpdatedAt);
+    expect(firstStore["agent:a:unrelated"]).toEqual(unrelatedEntry);
+    expect(secondStore["agent:b:other"]?.pluginExtensions).toBeUndefined();
+    expect(secondStore["agent:b:other"]?.updatedAt).toBeGreaterThan(beforeUpdatedAt);
+  });
 });
